p2/requirements.txt
numpy==1.19.2
pandas==1.1.3
scikit-learn==0.23.2
xgboost==1.2.1
lightgbm==3.0.0
joblib>=0.14.1
pytest==6.1.2
pytest-html==2.1.1
pytest-cov==2.10.1
hypothesis==5.41.1
pytest-xdist==2.1.0
python-coveralls==2.9.3
flake8==3.8.4
dvc==1.9.1
pyyaml==5.3.1
kedro==0.16.6